pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pediassist"
version = "0.1.0"
description = "AI-Powered Pediatric Subspecialty Treatment Assistant"
authors = [{name = "PediAssist Team", email = "team@pediassist.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Information Analysis",
]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "chromadb>=0.4.18",
    "litellm>=1.11.0",
    "openai>=1.3.0",
    "anthropic>=0.7.0",
    "tiktoken>=0.5.1",
    "click>=8.1.7",
    "rich>=13.7.0",
    "textual>=0.44.1",
    "pandas>=2.1.3",
    "numpy>=1.26.2",
    "sentence-transformers>=2.2.2",
    "cryptography>=41.0.7",
    "bcrypt>=4.1.1",
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.2",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "ruff>=0.1.6",
]

[project.scripts]
pediassist = "pediassist.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pediassist*"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
asyncio_mode = "auto"